requirements.txt
Django==4.2
css-inline==0.10.1
docutils==0.20.1
django-allauth==0.52.0
django-cleanup==7.0.0
django-compressor==4.3.1
django-debug-toolbar==4.0.0
django-libsass==0.9
django-phonenumber-field==7.1.0
django-simple-captcha==0.5.17
environs==9.5.0
gunicorn==20.1.0
libsass==0.22.0
mysqlclient==2.2.0
phonenumbers==8.13.13
Pillow==9.5.0
whitenoise==6.4.0
yookassa==2.3.6
premailer==3.10.0